JD-727: add aggregates [7]

+++ src/main/java/com/rbkmoney/fraudbusters/domain/dgraph/DgraphAggregates.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class DgraphAggregates {
 
     private int count;
-    private long sum;
+    private double sum;
     private DgraphMetrics metrics;
 
     public DgraphAggregates setQueryMetrics(DgraphMetrics metrics) {
